chore: isolate policy testing utilities
